Change to WC()->shipping()->calculate_shipping() for calculating shipping
Per @vedanshujain's [feedback](https://github.com/woocommerce/woocommerce/pull/30855#pullrequestreview-789313637)
we're calculating the packages using the "cart" and then using
WC()->shipping() to calculate shipping.

Note: This still uses the in-memory session instead of the cookie. We're
loading up an order that is stored in the database and calculating
shipping for that.